refactor: extracted component to its own file
